Convert non-Cloudwatch plugins to new read api, switch to ES6.
Modify read adapter to use new read api. Validation of options like
-from/-to/etc is mostly pushed into the base class. ReadAWS still
verifies that only the base options + possible -cloudwatch are present.

Remove all the explicit polling and working with the scheduler. Now,
every time read() is called all the plugins poll() to get points.

-limit is still ignored right now.

Most of the plugin interactions are unchanged--they still poll and
return arrays of metrics and events. One change is that they are given
a time to attach to all points (now_), which is the from value of the
call to read().

Convert all plugins to use ES6, specifically:
  - classes instead of extendable-base
  - use let whenever possible
  - use fat arrow functions to preserve this in function scopes
  - modify all array.forEach to for item of array

I noticed a few cases where I wasn't explicitly returning Promises at
the end of a chain, so clean that up.