Keep only Cloudwatch support.
Update READMEs to only mention Cloudwatch metrics and not the other
demographics, aggregates, or metrics.

Update README to mention new arguments -period, -statistics, -from, -to.

Rename files/objects with *aws* to *cloudwatch*.

Simplify the object created by parsing the filter ast. It's just a
mapping from product to list of item ids.

Remove all the other plugins, including the factory that created
objects from product names. Now the set of supported products is
simply in the read proc and given to the filter parser. The
CloudWatchMon object is now created explicitly instead of indirectly
through the factory.

AWS configuration moves back to the read proc now that the factory is gone.

Move configuring the CloudwatchMon object with the set of
products/items to monitor into the constructor, and specify it
directly instead of via function calls.

Simplify what's returned from poll() to just be a list of metrics.