Update to support new read api.
If not specified, from is set to 10 minutes ago and to is set to
now. This supports the common case for Cloudwatch, where the default
metrics are delayed by 5 minutes.

set periodicLiveRead() to true so someone can read starting at :now:
with -every. One downside is that in the case where default aws
metrics are delayed by 5 minutes, choosing an -every of less than 5
minutes will cause the read to miss everything, as the read will not
return anything and then advance time. Even when using CloudWatch
advanced metrics (which you have to pay for), metrics are not
aggregated more often than once per minute.

To address this, log a warning when the value of -every is less than 5
minutes and throw an error when the value is -every is less than 1
minute. There are new adapter config options
disable_every_{warnings,errors} that allow suppressing these
errors/warnings for users who are sure of their choice of -every.

Add unit tests for throwing errors/warnings for small -every values.

Change the basic 'read cloudwatch' test to not provide any options so
all metrics are fetched for the default timerange.