Address feedback from PR.
- Just pass an array of objects (product name, product item name, item
  ids) to the CloudWatchMon class instead of a hash.
- Only flatten the arrays from all the products once, right before sorting.
- While parsing the filter expression, share some repeated code for
  throwing filter-related errors in a single method
  throwUnsupportedFilter().
- Make the set of supported products a constant.
- Use the new format for specifying allowed options.

Also apply some feedback from the aws PR to this PR.

 - Validate the config for required properties when the adapter is loaded.
 - Clean up the steps for reading the config in the cloudwatch adapter.
 - Also fix lint.

Add detailed unit test that tests fetching cloudwatch metrics for a few cases:
 - defaults (all products, 60 second period, Average)
 - using a statistic of Minimum
 - Using multiple statistics (Minumum, Average)
 - Using a different sampling period

The tests check the returned points for the expected
fields (metric_tyoe, product, namespace, etc.) and check the values of
the properties (verifying the statistic name for a given product is
one of the ones we know about and that the units is the expected value).

While doing this, found some bugs:
 - period/statistics were not being set correctly.
 - initialize was misspelled
 - Some metrics were missing from the docs.
 - Sometimes the dimensions array will be empty, so skip those cases.

I also found it's easier to use the namespace as a way to get all
metrics for a given product than by specifying item names, so switch
to that.

Finally, add some debug logs that make it easier to follow the process.