Support metrics in filtering expressions.
While working on the example programs for AWS/Cloudwatch, I found that
the Cloudwatch reads were pretty slow, as they had to fetch all
metrics, even if the read was really only interested in a single
metric like CPUUtilization.

To address this, add support for metrics in filtering expessions in
addition to items. This required a rework of FilterCloudWatchCompiler
to move beyond just a simple list of product= or item= filters. The
filtering expression now supports using AND to combine a
product=<product>, item=<item>, and/or metric=<metric> selector
together into a single condition. You can still use a concise format
of item=<product>:<item> or metric=<product>:<metric>, which is
essentially shorthand for product=<product> AND item=<item>. OR is
used to build up a list of conditions.

Within CloudWatchMon, it's given a list of conditions instead of a
list of products, as a product may appear in multiple conditions. If a
condition has a list of metrics, it is used to filter down the result
from listMetrics to only the desired metrics before calling
getMetricStatistics.

Add a bunch of unit tests that test all the new negative and positive
cases this introduces. While doing this, I found that some of the
negative tests were not actually working. In cases where the incorrect
filtering expression falsely compiled successfully, it wouldn't throw
an error at all and as a result wouldn't get to the expects() in the
catch handler. Fix that by throwing an error if compile doesn't.